requirements-tests.txt
click==8.1.8
cryptography
django-stubs
freezegun
ninja-schema>=0.14.1
pytest
pytest-asyncio==0.24.0
pytest-cov
pytest-django
python-jose==3.3.0
ruff ==0.11.2
